Version release: 2.0.0
This release isn't an improvement. It makes it barely work with
Tiddlywiki V5.2.x, because due to changes in the new Tiddlywiki, there
is no longer a good way to uglify the javascript tiddlers. This new
version is slow, it's caching doesn't work, and it can be easily
broken with future Tiddlywiki changes.